requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.29
aiosqlite==0.20.0
python-dotenv==1.0.1
google-genai>=1.0.0
livekit-agents~=1.4
websockets>=13.0
python-dateutil>=2.9.0
